args 내부 데이터를 분리해서 전달 받도록 수정
