+++ shared/format_util.py
@@ -3,6 +3,7 @@ from typing import List
 
 from app.cognitive_service.agent_tool.calendar_tool import TravelPlansInput
 from app.models.travel_plan import TravelPlan
+from app.schemes.calendar_scheme import TravelPlanInput
 
 
 def format_user_messages_with_index(messages):
@@ -25,18 +26,17 @@ def to_html_format(html_body):
 
 
 
-def convert_input_to_travel_plans(inputs: TravelPlansInput) -> tuple[str, List[TravelPlan]]:
+def convert_input_to_travel_plans(session_id: str, inputs: List[TravelPlanInput]) -> list[TravelPlan]:
     if not inputs:
         raise ValueError("입력 일정 정보가 비어 있습니다.")
 
-    session_id = inputs.session_id
     travel_plans = [
         TravelPlan(
             session_id=session_id,
-            trip_date=plan.trip_date,
-            trip_schedule=plan.trip_schedule,
+            trip_date=plan_input.trip_date,
+            trip_schedule=plan_input.trip_schedule,
             created_at=date.today().isoformat()
         )
-        for plan in inputs.plans
+        for plan_input in inputs
     ]
-    return session_id, travel_plans
+    return travel_plans
